+++ src/main/java/com/github/aksc/InputOutput/Parser.java
@@ -1,7 +1,7 @@
 package com.github.aksc.InputOutput;
 
 import com.github.aksc.DerivationSystem;
-import com.github.aksc.Exceptions.BadInputException;
+import com.github.aksc.ErrorHandling.BadInputException;
 import com.github.aksc.Grammar.Symbol;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
