+++ src/main/java/com/github/aksc/Main.java
@@ -1,7 +1,7 @@
 package com.github.aksc;
 
-import com.github.aksc.Exceptions.BadInputException;
-import com.github.aksc.Exceptions.BadLanguageException;
+import com.github.aksc.ErrorHandling.BadInputException;
+import com.github.aksc.ErrorHandling.BadLanguageException;
 import com.github.aksc.Grammar.Symbol;
 import com.github.aksc.InputOutput.CommandLineInput;
 import com.github.aksc.InputOutput.CommandLineOutput;
